# openapi.yaml
openapi: 3.0.1
info:
  description: Documentation API v1.0
  title: 'API сервер'
  version: "1.0"
servers:
- description: Generated server url
  url: http://localhost:8080
...
paths:
  /admin/categories:
    post:
      description: "Обратите внимание: имя категории должно быть уникальным"
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCategoryDto'
        description: данные добавляемой категории
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
          description: Категория добавлена
        "400":
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: "Field: name. Error: must not be blank. Value: null"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: "could not execute statement; SQL [n/a]; constraint [uq_category_name];\
                  \ nested exception is org.hibernate.exception.ConstraintViolationException:\
                  \ could not execute statement"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Добавление новой категории
      tags:
      - "Admin: Категории"
      x-content-type: application/json
      x-accepts:
      - application/json
  /admin/categories/{catId}:
    delete:
      description: "Обратите внимание: с категорией не должно быть связано ни одно\
        го события."
      operationId: deleteCategory
      parameters:
      - description: id категории
        explode: false
        in: path
        name: catId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Категория удалена
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Category with id=27 was not found
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Категория не найдена или недоступна
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: For the requested operation the conditions are not met.
                message: The category is not empty
                timestamp: 2023-01-21 16:56:19
              schema:
                $ref: '#/components/schemas/ApiError'
          description: "Существуют события, связанные с категорией"
      summary: Удаление категории
      tags:
      - "Admin: Категории"
      x-accepts:
      - application/json
    patch:
      description: "Обратите внимание: имя категории должно быть уникальным"
      operationId: updateCategory
      parameters:
      - description: id категории
        explode: false
        in: path
        name: catId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        description: Данные категории для изменения
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
          description: Данные категории изменены
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Category with id=27 was not found
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Категория не найдена или недоступна
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: "could not execute statement; SQL [n/a]; constraint [uq_category_name];\
                  \ nested exception is org.hibernate.exception.ConstraintViolationException:\
                  \ could not execute statement"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Изменение категории
      tags:
      - "Admin: Категории"
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /admin/users:
    get:
      description: |-
        Возвращает информацию обо всех пользователях (учитываются параметры ограничения выборки), либо о конкретных (учитываются указанные идентификаторы)

        В случае, если по заданным фильтрам не найдено ни одного пользователя, возвращает пустой список
      operationId: getUsers
      parameters:
      - description: id пользователей
        explode: true
        in: query
        name: ids
        required: false
        schema:
          items:
            format: int64
            type: integer
          type: array
        style: form
      - description: "количество элементов, которые нужно пропустить для формирова\
          ния текущего набора"
        explode: true
        in: query
        name: from
        required: false
        schema:
          default: 0
          format: int32
          minimum: 0
          type: integer
        style: form
      - description: количество элементов в наборе
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserDto'
                type: array
          description: Пользователи найдены
        "400":
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: "Failed to convert value of type java.lang.String to required\
                  \ type int; nested exception is java.lang.NumberFormatException:\
                  \ For input string: ad"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Получение информации о пользователях
      tags:
      - "Admin: Пользователи"
      x-accepts:
      - application/json
    post:
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserRequest'
        description: Данные добавляемого пользователя
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
          description: Пользователь зарегистрирован
        "400":
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: "Field: name. Error: must not be blank. Value: null"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: "could not execute statement; SQL [n/a]; constraint [uq_email];\
                  \ nested exception is org.hibernate.exception.ConstraintViolationException:\
                  \ could not execute statement"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Добавление нового пользователя
      tags:
      - "Admin: Пользователи"
      x-content-type: application/json
      x-accepts:
      - application/json
  /admin/users/{userId}:
    delete:
      operationId: delete
      parameters:
      - description: id пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Пользователь удален
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: User with id=555 was not found
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Пользователь не найден или недоступен
      summary: Удаление пользователя
      tags:
      - "Admin: Пользователи"
      x-accepts:
      - application/json
  /categories:
    get:
      description: "В случае, если по заданным фильтрам не найдено ни одной катего\
        рии, возвращает пустой список"
      operationId: getCategories
      parameters:
      - description: "количество категорий, которые нужно пропустить для формирова\
          ния текущего набора"
        explode: true
        in: query
        name: from
        required: false
        schema:
          default: 0
          format: int32
          minimum: 0
          type: integer
        style: form
      - description: количество категорий в наборе
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoryDto'
                type: array
          description: Категории найдены
        "400":
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: "Failed to convert value of type java.lang.String to required\
                  \ type int; nested exception is java.lang.NumberFormatException:\
                  \ For input string: ad"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Получение категорий
      tags:
      - "Public: Категории"
      x-accepts:
      - application/json
  /categories/{catId}:
    get:
      description: "В случае, если категории с заданным id не найдено, возвращает\
        \ статус код 404"
      operationId: getCategory
      parameters:
      - description: id категории
        explode: false
        in: path
        name: catId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
          description: Категория найдена
        "400":
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: "Failed to convert value of type java.lang.String to required\
                  \ type long; nested exception is java.lang.NumberFormatException:\
                  \ For input string: ad"
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Category with id=19 was not found
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Категория не найдена или недоступна
      summary: Получение информации о категории по её идентификатору
      tags:
      - "Public: Категории"
      x-accepts:
      - application/json
...
components:
  schemas:
    ApiError:
      description: Сведения об ошибке
      example:
        reason: For the requested operation the conditions are not met.
        message: Only pending or canceled events can be changed
        errors: []
        status: FORBIDDEN
        timestamp: 2022-06-09 06:27:23
      properties:
        errors:
          description: Список стектрейсов или описания ошибок
          example: []
          items:
            description: Список стектрейсов или описания ошибок
            example: "[]"
            type: string
          type: array
        message:
          description: Сообщение об ошибке
          example: Only pending or canceled events can be changed
          type: string
        reason:
          description: Общее описание причины ошибки
          example: For the requested operation the conditions are not met.
          type: string
        status:
          description: Код статуса HTTP-ответа
          enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: FORBIDDEN
          type: string
        timestamp:
          description: Дата и время когда произошла ошибка (в формате "yyyy-MM-dd
            HH:mm:ss")
          example: 2022-06-09 06:27:23
          type: string
      type: object
    CategoryDto:
      description: Категория
      example:
        name: Концерты
        id: 1
      properties:
        id:
          description: Идентификатор категории
          example: 1
          format: int64
          readOnly: true
          type: integer
        name:
          description: Название категории
          example: Концерты
          maxLength: 50
          minLength: 1
          type: string
      required:
      - name
      type: object
...
    NewCategoryDto:
      description: Данные для добавления новой категории
      example:
        name: Концерты
      properties:
        name:
          description: Название категории
          example: Концерты
          maxLength: 50
          minLength: 1
          type: string
      required:
      - name
      type: object
...
    NewUserRequest:
      description: Данные нового пользователя
      example:
        name: Иван Петров
        email: ivan.petrov@mail.ru
      properties:
        email:
          description: Почтовый адрес
          example: ivan.petrov@mail.ru
          maxLength: 254
          minLength: 6
          type: string
        name:
          description: Имя
          example: Иван Петров
          maxLength: 250
          minLength: 2
          type: string
      required:
      - email
      - name
      type: object
...
    UserDto:
      description: Пользователь
      example:
        name: Петров Иван
        id: 1
        email: petrov.i@mail.ru
      properties:
        email:
          description: Почтовый адрес
          example: petrov.i@mail.ru
          type: string
        id:
          description: Идентификатор
          example: 1
          format: int64
          readOnly: true
          type: integer
        name:
          description: Имя
          example: Петров Иван
          type: string
      required:
      - email
      - name
      type: object
...